Client utils moved to clinet folder.
