+++ src/client/app/components/UI/widgets/widgetBiggestCategories.jsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Widget from '../../common/widget';
-import { greenColor, redColor } from '../../../../../constants/colors';
-import showElement from '../../../../../utils/console/showElement';
 import { useSelector } from 'react-redux';
-import { getCategoriesList, getCategoryById } from '../../../store/categories';
-import { getIconById, getIconsList } from '../../../store/icons';
-import SVGIcon from '../../common/svgIcon';
+import { getCategoriesList } from '../../../store/categories';
+import { getIconsList } from '../../../store/icons';
 import { operationPropType } from '../../../../../types/propTypes';
 import StatisticPlate from '../../common/statisticsPlate';
 import roundToHundredths from '../../../../../utils/math/roundToHundredths';
