+++ src/client/app/components/UI/widgets/widgetBiggestOperations.jsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Widget from '../../common/widget';
-import { greenColor, redColor } from '../../../../../constants/colors';
-import showElement from '../../../../../utils/console/showElement';
 import { useSelector } from 'react-redux';
 import { getCategoryById } from '../../../store/categories';
 import { getIconById } from '../../../store/icons';
-import SVGIcon from '../../common/svgIcon';
 import { operationPropType } from '../../../../../types/propTypes';
 import StatisticPlate from '../../common/statisticsPlate';
