+++ src/client/app/components/common/modalWindow.jsx
@@ -2,10 +2,10 @@ import React, { cloneElement, useCallback, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { nodesPropType } from '../../../../types/propTypes';
 import { mainColor } from '../../../../constants/colors';
-import closeModalWindow from '../../../../utils/modals/closeModalWindow';
+import closeModalWindow from '../../utils/modals/closeModalWindow';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import showElement from '../../../../utils/console/showElement';
-import openModal from '../../../../utils/modals/openModal';
+import showElement from '../../utils/console/showElement';
+import openModal from '../../utils/modals/openModal';
 
 export default function ModalWindow({
   headTitle = 'Введите данные',
