+++ src/client/app/services/operations.service.ts
@@ -1,7 +1,7 @@
 import { Operation } from "../../../types/types";
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import showElement from "../../../utils/console/showElement";
-import displayError from "../../../utils/errors/onClient/displayError";
+import displayError from "../utils/errors/onClient/displayError";
 import { createCRUDServiceFunctions } from "./crud.service";
 import httpService from "./http.service";
