+++ src/client/app/utils/calculateAmount.ts
