+++ src/client/app/utils/capitalize.ts
@@ -1,0 +1,3 @@
+export default function capitalize(string:string) {
+  return string.replace(string[0], string[0].toUpperCase());
+}
