+++ src/client/app/utils/console/showError.ts
@@ -1,0 +1,3 @@
+export default function showError(message:unknown) {
+  console.error(`ERROR OCCURRED: `, message);
+}
