+++ src/client/app/utils/errors/onClient/displayError.ts
