+++ src/client/app/utils/getCategoryIcon.ts
