+++ src/client/app/utils/math/roundToHundredths.ts
@@ -1,0 +1,3 @@
+export default function roundToHundredths(number: number): number {
+  return Math.round(number * 100) / 100;
+}
