+++ src/client/app/utils/modals/openModal.ts
@@ -1,0 +1,15 @@
+export default function openModal(modal: HTMLDialogElement) {
+  if (!modal) return;
+
+  modal.showModal();
+  // document.addEventListener('click', closeModalOnBackdropClick);
+
+  // function closeModalOnBackdropClick(event: Event) {
+  //   if (event.target instanceof Element) {
+  //     const { target } = event;
+  //     if (target.nodeName === 'DIALOG') {
+  //       closeModalWindow(modal);
+  //     }
+  //   }
+  // }
+}
