+++ src/client/app/utils/validation/flashInvalidInputs.ts
@@ -1,5 +1,3 @@
-import showElement from "./console/showElement.ts";
-
 export default function flashInvalidInputs(...inputs: HTMLElement[]) {
   inputs.forEach((input) => {
     if (input.nodeName === 'DIV') {
