+++ src/client/app/utils/validation/flashOffInvalidInputs.ts
@@ -1,0 +1,8 @@
+export default function flashOffInvalidInputs(...inputs: HTMLElement[]) {
+  inputs.forEach((input) => {
+    if (input.nodeName === 'DIV') {
+      input.classList.remove('bordered');
+    }
+      input.classList.remove('invalid');
+  });
+}
