+++ src/utils/console/showError.ts
@@ -1,6 +1,4 @@
-import { redLog } from "./coloredLogs.ts";
-
 export default function showError(message:unknown) {
-  redLog(`ERROR OCCURRED: `);
-  console.log(message);
+  console.error(`ERROR OCCURRED: `);
+  console.error(message);
 }
