--- src/utils/date/getMonthBorders.ts
@@ -1,8 +1,0 @@
-export default function getMonthBorders() {
-  const today = new Date();
-
-  const firstDay = new Date(today.getFullYear(), today.getMonth(), 1);
-  const lastDay = new Date(today.getFullYear(), today.getMonth() + 1, 0);
-
-  return { firstDay, lastDay };
-}
