--- src/utils/date/getTodayDate.ts
@@ -1,6 +1,0 @@
-export default function getTodayDate(): Date {
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-
-  return today;
-}
