--- src/utils/flashOffInvalidInputs.ts
@@ -1,8 +1,0 @@
-export default function flashOffInvalidInputs(...inputs: HTMLElement[]) {
-  inputs.forEach((input) => {
-    if (input.nodeName === 'DIV') {
-      input.classList.remove('bordered');
-    }
-      input.classList.remove('invalid');
-  });
-}
