--- src/utils/modals/openModal.ts
@@ -1,18 +1,0 @@
-import showElement from "../console/showElement.ts";
-import closeModalWindow from "./closeModalWindow.ts";
-
-export default function openModal(modal: HTMLDialogElement) {
-  if (!modal) return;
-
-  modal.showModal();
-  // document.addEventListener('click', closeModalOnBackdropClick);
-
-  // function closeModalOnBackdropClick(event: Event) {
-  //   if (event.target instanceof Element) {
-  //     const { target } = event;
-  //     if (target.nodeName === 'DIALOG') {
-  //       closeModalWindow(modal);
-  //     }
-  //   }
-  // }
-}
