--- src/utils/modals/showErrorMessage.js
@@ -1,6 +1,0 @@
-export default function showErrorMessage(window) {
-  window.removeAttribute('hidden');
-  setTimeout(() => {
-    window.setAttribute('hidden', '');
-  }, 5000);
-}
